Update the `FfmpegFormat.Reader._terminate()` method to deal with webcam (#346)
* add rudimentary test for ffmpeg process termination when using webcam

* when streaming from webcam, terminate ffmpeg process using communicate('q') instead of kill() (fixes #343)

* fix print statement in webcam test

* add psutil to travis yaml and add need_internet() to ffmpeg webcam termination test

* add psutil to appveyor.yml as well...

+++ imageio/plugins/ffmpeg.py
@@ -501,7 +501,11 @@ class FfmpegFormat(Format):
             # Terminate process
             # Using kill since self._proc.terminate() does not seem
             # to work for ffmpeg, leaves processes hanging
-            self._proc.kill()
+            if self.request._video:
+                # When streaming from webcam, quit by sending 'q' to ffmpeg
+                self._proc.communicate('q')
+            else:
+                self._proc.kill()
             
             # Tell threads to stop when they have a chance. They are probably
             # blocked on reading from their file, but let's play it safe.
